chore: Extract warning-suppression macros to doctest/parts/warnings.h
